move beanstalkd to top level project
