+++ python/python/bystro/search/utils/messages.py
@@ -1,53 +1,5 @@
-from enum import Enum
-from typing import get_type_hints
-
-from msgspec import Struct, field
-
+from bystro.beanstalkd.messages import BaseMessage, Struct
 from bystro.search.utils.annotation import AnnotationOutputs
-
-SubmissionID = str | int
-BeanstalkJobID = int
-
-
-class BaseMessage(Struct, frozen=True):
-    submissionID: SubmissionID
-
-    @classmethod
-    def keys_with_types(cls) -> dict:
-        return get_type_hints(cls)
-
-
-class FailedJobMessage(Struct, frozen=True):
-    submissionID: SubmissionID
-    reason: str
-
-
-class InvalidJobMessage(Struct, frozen=True):
-    # Invalid jobs that are invalid because the submission breaks serialization invariants
-    # will not have a submissionID as that ID is held in the serialized data
-    queueID: BeanstalkJobID
-    reason: str
-
-
-class Event(str, Enum):
-    """Beanstalkd Event"""
-
-    PROGRESS = "progress"
-    FAILED = "failed"
-    STARTED = "started"
-    COMPLETED = "completed"
-
-
-class ProgressData(Struct):
-    progress: int = 0
-    skipped: int = 0
-
-
-class ProgressMessage(BaseMessage, frozen=True):
-    """Beanstalkd Message"""
-
-    event: str = Event.PROGRESS
-    data: ProgressData = field(default_factory=ProgressData)
 
 
 class IndexJobData(BaseMessage, frozen=True):
